Move an array test
